ethdev: add mhpsdp hardware port match item
For the multiple hardware ports connect to a single DPDK port (mhpsdp),
currently there is no information to indicate the packet belongs to
which hardware port.

This patch introduces a new hardware port item in rte flow API, and
the port value reflects the hardware port of the received packets.

While uses the hardware port as a matching item in the flow, and sets the
same port value on the tx queue, then the packet can be sent from the same
hardware port with received.

This patch also adds the testpmd command line to match the new item:
	flow create 0 ingress group 0 pattern mhpsdp_hw_port hwport is 1 /
	end actions queue index 0 / end

The above command means that creates a flow on a single DPDK port and
matches the packet from the first physical port (the hwport 1
stands for the first port) and redirects these packets into RxQ 0.

Signed-off-by: Jiawei Wang <jiaweiw@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>