net/hns3: refactor set RSS hash algorithm and key interface
The hns3_rss_set_algo_key() is used to set RSS hash algorithm and key to
hardware. The maximum times of command sent to firmware depend on the
length of key. However, now this times is fixed, which isn't good for
key expansion. In addition, hash algorithm comes from rss_info::hash_algo
maintained in driver, which also isn't good for the usage of this function.
This patch has to use hash algorithm and key length as the input parameters
of this interface.

Fixes: c37ca66f2b27 ("net/hns3: support RSS")
Cc: stable@dpdk.org

Signed-off-by: Huisong Li <lihuisong@huawei.com>
Signed-off-by: Dongdong Liu <liudongdong3@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>