drivers: remove compile-time option for IEEE 1588
The option RTE_LIBRTE_IEEE1588 has no effect on any library
unlike its name.

Also we are suppose to enable/disable features dynamically,
not at compilation time.

And the best is that this macro is neither documented,
nor in rte_config.h.

It looks to be a mistake keeping this flag, so it is removed,
meaning always enabled.
PS: it is disabling vector paths of some drivers.

Signed-off-by: Thomas Monjalon <thomas@monjalon.net>
Signed-off-by: 0-day Robot <robot@bytheb.org>