net/iavf: remove avx512 specific Rx queue rearm code
'iavf_rxq_rearm' in avx512 path is optimized to improve the performance.
But after the commit a2833ecc5ea4 ("mempool: fix get objects from mempool
with cache"), this avx512 specific optimization is not necessary.
This patch remove the unnecessary PMD specific optimization to make the
code easier to maintain and get the benefit from the enhancement of common
lib.

Reported-by: Haijun Chu <haijun.chu@intel.com>
Signed-off-by: Wenzhuo Lu <wenzhuo.lu@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>