gso: remove logtype
If a large packet is passed into GSO routines of unknown protocol
then library would log a message and pass it through. This is incorrect
behaviour on many levels:
  - it allows oversize packet to get passed on to NIC driver
  - no direct return is visible to applications
  - if it happens once, many more will follow and log will fill.
  - bonus it is only log message with GSO type.

The fix is to just return -EINVAL which is what this library
does in many other places when looking at headers.

Fixes: 119583797b6a ("gso: support TCP/IPv4 GSO")
Cc: jiayu.hu@intel.com
Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>