common/cnxk: support IPv6 field hash in RTE Flow
CN103xx has the field hash feature which can hash 128 bit values
to 32 bit values. The hashed value can be used in MCAM keys thereby
freeing up space in the key. This patch adds support to make use of
this hardware feature to hash IPv6 addresses and put them in
the MCAM keys.

Signed-off-by: Satheesh Paul <psatheesh@marvell.com>
Reviewed-by: Kiran Kumar K <kirankumark@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>